# [WIP] Implement cache mechanism for GMT date strings.