This improves performance about 5~8%.

## Benchmarks

My environment.

```console
$ uname -mrv
4.4.0-98-generic #121-Ubuntu SMP Tue Oct 10 14:24:03 UTC 2017 x86_64
$ cat /proc/cpuinfo| grep 'model name'| head -1
model name      : Intel(R) Core(TM) i7-7500U CPU @ 2.70GHz
$ dmd --version
DMD64 D Compiler v2.077.1
Copyright (c) 1999-2017 by Digital Mars written by Walter Bright
```

### master (head: 48a9ddc2e97245f7356f448a8a29912996f34e35)

```console
$ wrk -c 1000 -d 10s http://localhost:8080/
Running 10s test @ http://localhost:8080/
  2 threads and 1000 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    16.48ms   17.86ms 483.84ms   99.03%
    Req/Sec    31.07k     4.00k   50.20k    75.38%
  615503 requests in 10.09s, 99.20MB read
Requests/sec:  60976.43
Transfer/sec:      9.83MB
wrk -c 1000 -d 10s http://localhost:8080/  0.88s user 3.42s system 42% cpu 10.103 total
$ wrk -c 1000 -d 10s http://localhost:8080/
Running 10s test @ http://localhost:8080/
  2 threads and 1000 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    15.94ms   17.42ms 490.98ms   99.31%
    Req/Sec    31.88k     2.20k   35.40k    71.00%
  634679 requests in 10.08s, 102.29MB read
Requests/sec:  62983.25
Transfer/sec:     10.15MB
wrk -c 1000 -d 10s http://localhost:8080/  0.82s user 3.40s system 41% cpu 10.087 total
$ wrk -c 1000 -d 10s http://localhost:8080/
Running 10s test @ http://localhost:8080/
  2 threads and 1000 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    16.48ms   20.05ms 504.62ms   99.19%
    Req/Sec    31.70k     2.49k   40.94k    75.00%
  630867 requests in 10.09s, 101.68MB read
Requests/sec:  62497.57
Transfer/sec:     10.07MB
wrk -c 1000 -d 10s http://localhost:8080/  0.92s user 3.44s system 43% cpu 10.107 total
```

### date (head: 75454b3c406207151a4ef03461adf660bb893f5d)

```console
$ wrk -c 1000 -d 10s http://localhost:8080/
Running 10s test @ http://localhost:8080/
  2 threads and 1000 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    15.16ms   12.94ms 454.33ms   99.43%
    Req/Sec    33.02k     1.87k   37.29k    66.00%
  657144 requests in 10.09s, 105.91MB read
Requests/sec:  65134.60
Transfer/sec:     10.50MB
wrk -c 1000 -d 10s http://localhost:8080/  0.98s user 3.87s system 47% cpu 10.100 total
$ wrk -c 1000 -d 10s http://localhost:8080/
Running 10s test @ http://localhost:8080/
  2 threads and 1000 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    16.15ms   21.33ms 523.13ms   99.19%
    Req/Sec    32.78k     1.65k   36.31k    72.50%
  652490 requests in 10.08s, 105.16MB read
Requests/sec:  64740.40
Transfer/sec:     10.43MB
wrk -c 1000 -d 10s http://localhost:8080/  0.98s user 4.06s system 49% cpu 10.091 total
$ wrk -c 1000 -d 10s http://localhost:8080/
Running 10s test @ http://localhost:8080/
  2 threads and 1000 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    15.26ms   13.76ms 447.96ms   99.16%
    Req/Sec    32.91k     2.29k   38.67k    78.50%
  654998 requests in 10.09s, 105.57MB read
Requests/sec:  64915.08
Transfer/sec:     10.46MB
wrk -c 1000 -d 10s http://localhost:8080/  1.11s user 3.92s system 49% cpu 10.105 total
```