@s-ludwig 2017-12-20
Wow, I'm surprised that this has such an effect, I never saw this in a profiler output (in constrast to two other factors that have a huge optimization potential). Two things that are currently still wrong in the changed code: The code uses `Clock.currTime()` (which, if called without an `UTC()` argument, actually used to be rather expensive) instead of the `time` argument, and the result is cached potentially until after the life time of the underlying allocator, which is logically created and destroyed per request, so that it could lead to use-after-free errors.

Maybe the speedup is simply based on the fact that it saves the allocator calls (should be cheap unless the first memory pool gets exhausted)? In that case another possibility would be to use a `FixedAppender` stored in `HTTPServerResponse`, like it's done with `m_contentLengthBuffer` in `HTTPClientRequest`.