@kubo39 2017-12-20
Thank you for your review!

> The code uses Clock.currTime() (which, if called without an UTC() argument, actually used to be rather expensive) instead of the time argument, and the result is cached potentially until after the life time of the underlying allocator, which is logically created and destroyed per request, so that it could lead to use-after-free errors.

Exactly. I think that it seems better to call `clock_gettime` directly, without `SysTime`.

 >  In that case another possibility would be to use a FixedAppender stored in HTTPServerResponse, like it's done with m_contentLengthBuffer in HTTPClientRequest.

Using `FixedAppender!(string, 29)` seems good, I want to investigate more.