@kubo39 2019-07-26
Old LDC with libevent or libasync causes segmentation fault, I'm not sure why.