@kubo39 2019-09-07
Update benchmark result.